requirements.txt
python-dotenv==0.14.0
requests==2.22.0